backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
groq
python-dotenv